Entry-Exit FD
Introduced Exit and Entry of final demand firms. A number of fixes and corrections have been introduced as necessary

Bank is recapitalised by government in case of negative capital. The gov. debt is increased and the (negative) bank capital is reset to 0.

FD firms set their price adjusting the lagged value. In case the previous time the price was increased because of extra price (demand higher than available stock and production), the return to standard level is slow. Parameter shiftExPrice determines the speed of adjustment.

GovDemTot (the demand by government) is re-written. It now targets the ratio of debt on GDP and unemployment.